+++ R/RepeatCircular.R
@@ -1,16 +1,14 @@
-#' Repeat minimum longitude
+#' Extend periodic data to any range
 #'
-#' ggplot's \code{\link[ggplot2]{coord_polar}} doesn't close contours in
-#' cyclical data so the workaround is to copy the leftmost data to the right.
+#' Periodic data can be defined only in one period and be extended to any arbitrary
+#' range.
 #'
 #' @param x a data.frame
 #' @param circular the name of the circular dimension
-#' @param max the value of the longitude-like column that the duplicated
-#' values should have. If `NULL`, defaults to max + resolution.
+#' @param range the range for the data to be extended to
 #'
 #' @return
-#' A data.frame with the values corresponding to the minimum longitude
-#' repeated next to the maximum.
+#' A data.frame.
 #'
 #' @examples
 #' library(ggplot2)
@@ -25,31 +23,59 @@
 #' g
 #'
 #' # But using RepeatCircular solves it.
-#' g %+% RepeatCircular(geopotential[date == date[1]])
+#' g %+% ExtendCircular(geopotential[date == date[1]], "lon", c(0, 360))
+#'
+#' # Aditionally data can be just repeatet to the right and
+#' # left
+#' ggplot(ExtendCircular(geopotential[date == date[1]], range = c(-180, 360 + 180)),
+#'        aes(lon, lat)) +
+#'     geom_contour(aes(z = gh))
 #'
 #' # The same behaviour is now implemented directly in geom_contour2
+#' # and geom_contour_fill
 #' ggplot(geopotential[date == date[1]], aes(lon, lat)) +
 #'     geom_contour2(aes(z = gh), circular = "x") +
-#'     coord_polar() +
-#'     ylim(c(-90, -10))
+#'     xlim(c(-180, 360+180))
 #'
 #' @seealso geom_contour2
 #'
 #' @family ggplot2 helpers
 #' @export
 #' @import data.table
-RepeatCircular <- function(x, circular = "lon", max = NULL) {
-    dt <- data.table::is.data.table(x)
-    data.table::setDT(x)
-    minlon <- x[, min(get(circular))]
-    border <- x[get(circular) == minlon, ]
-    max <- max(x[[circular]]) + ggplot2::resolution(x[[circular]])
-    border[, c(circular) := max]
-    full <- rbind(x, border)
-    if (dt == TRUE) {
-        return(full)
-    } else {
-        return(setDF(full))
-    }
+ExtendCircular <- function(x, circular = "lon", range = c(0, 360)) {
+    setDT(x)
+    res <- ggplot2::resolution(x[[circular]])
+    m <- min(x[[circular]])
+    M <- max(x[[circular]])
+
+    # How many steps form the left and right extremes
+    # represent the range
+    right <- trunc((max(range) - M)/res)
+    left <- trunc((min(range) - m)/res)
+
+    # New "grid"
+    x.new <- seq(m + left*res, M + right*res, by = res)
+
+    right <- right + uniqueN(x[[circular]]) - 1
+
+    # The old coord of the new grid
+    index <- seq(left, right)
+    index <- index %% length(unique(x[[circular]])) + 1
+    x.old <- unique(x[[circular]])[index]
+
+    x.new <- data.table(x.old, x.new)
+    colnames(x.new) <- c(circular, paste0(circular, "new"))
+
+    # Add values according to the old grid and then
+    # remove it.
+    y <- x[x.new, on = circular, allow.cartesian = TRUE]
+    set(y, NULL, circular, NULL )
+    setnames(y, paste0(circular, "new"), circular)
+    return(y)
 }
 
+#' @export
+#' @rdname ExtendCircular
+RepeatCircular <- function(x, circular = "lon", max = NULL) {
+    .Deprecated("ExtendCircular")
+}